Further fixes to avoid rounding issues

+++ src/main/java/com/github/ingaelsta/salarycalculator/entity/Salary.java
@@ -3,15 +3,17 @@ package com.github.ingaelsta.salarycalculator.entity;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 @AllArgsConstructor
 public class Salary {
     private Long employeeId;
     private Integer month;
     private Integer year;
-    private Double baseSalary;
-    private Double calculatedNonTaxableSum;
-    private Double calculatedIncomeTax;
-    private Double calculatedSocialTax;
-    private Double calculatedPayout;
+    private BigDecimal baseSalary;
+    private BigDecimal calculatedNonTaxableSum;
+    private BigDecimal calculatedIncomeTax;
+    private BigDecimal calculatedSocialTax;
+    private BigDecimal calculatedPayout;
 }
