+++ src/main/resources/schema.sql
@@ -2,7 +2,7 @@ create table employee (
     id bigint not null,
     name varchar(255) not null,
     surname varchar(255) not null,
-    base_Salary double not null,
+    base_Salary decimal(10, 2) not null,
     dependants int,
     use_Non_Taxable_Minimum bit,
     primary key (id)
@@ -11,7 +11,7 @@ create table employee (
 create table constant (
     id bigint not null AUTO_INCREMENT,
     name varchar(255) not null,
-    val double not null,
+    val decimal(10, 4) not null,
     start_Date date,
     end_Date date,
     primary key (id),
